+++ NEWS.md
@@ -41,6 +41,8 @@ There is a possibility to get package dependency versions used for the workflows
  
 - The function `dplyr::transmute()` is superseded in favor of `dplyr::mutate(.keep = "none")`. Consequently, all the admiral functions that utilized the former have been updated accordingly. (#2274)
 
+- The templates for ADPP and ADPC are updated for missing variables (#2308) and to make `ATPT` and `ATPTN` consistent. (#2328)
+
 # admiral 1.0.1
 
 - Fix bug in `derive_vars_query()` where if AE terms were in mixed case no terms are flagged. (#2311)