+++ NEWS.md
@@ -1,3 +1,7 @@
+# admiral 1.0.2
+
+- Fix bug in `derive_param_tte()` where argument `dataset` populated and argument `by_vars` is dynamic. (#2336)
+
 # admiral 1.0.1
 
 - Fix bug in `derive_vars_query()` where if AE terms were in mixed case no terms are flagged. (#2311)